Many Improvements! Light Show!
Transparency now works
Cleaned out fragment shader
deltaTimes (timesteps) are now floats counted in seconds
inputComponents now self bind
particles self update
cleaned up TestGame

+++ Thomas/ThomasTheGameEngine/TestGame/TestGame.h
@@ -22,7 +22,7 @@ public:
 protected:
 
 	/* Game Update Code Here */
-	void Update(Uint32 _timestep);
+	void Update(float _timestep);
 
 	/* Game Render Code Here */
 	void Render();
